simplified count descriptions
Removed all the superfluous "Alphabetical index of the ... " as it just looks silly. First of all, the list better be alphabetical otherwise there would be a bug, the whole point is to get a sorted list. Second, there's no need to use such theatrical descriptions for something as simple as a count.