differentiate between series info and other books in series in book detail
Added a new localization entry for book details for the link to the other books in the same series because on the feed output this gets used as title on some ereaders. And some (e.g. Stanza) place it at a location with already limited space, causing most of the important bit to be clipped by whatever else is on the same line. Contrary to for instance Easy Search, which pulls a completely different title from who knows where and displays just the Author Name(s) or the Series name depending on which of the two links you select.