Update OPDS_renderer.php
HTML should not be written into the XML stream as they are both tag based languages.  Depending on the complexity of the rich text in Calibre that populates the content field, this can stop validation and break some readers.  Notable PocketReader on Android which will not process an invalid OPDS.

4.1.3.3.  Processing Model
   2.  If the value of "type" is "html", the content of atom:content
       MUST NOT contain child elements and SHOULD be suitable for
       handling as HTML [HTML].  The HTML markup MUST be escaped; for
       example, "<br>" as "&lt;br>".  The HTML markup SHOULD be such
       that it could validly appear directly within an HTML <DIV>
       element.  Atom Processors that display the content MAY use the
       markup to aid in displaying it.