Update fetch.php to lower memory consumption
On big eBooks like comic books you can hit the PHP memory limit. With this changes there is no need to raise the memory limit in php.ini. Also readfile is recommended in the php documentation instead of fpassthru for most cases so I changed it. Also if you download binary data make sure that there is no output after that therefore an explicit exit() at the end.